Revert "fix uint int comparisons in sound.cpp"
This reverts commit 193528fb58423dfc2fb38e8f02f6163a3a8d9e51.

+++ src/engine/interface/sound.cpp
@@ -54,7 +54,7 @@ struct SoundConfig
 
     bool hasslot(const soundslot *p, const std::vector<soundslot> &v) const
     {
-        return p >= v.data() + slots && p < v.data() + slots+numslots && slots+numslots < static_cast<int>(v.size());
+        return p >= v.data() + slots && p < v.data() + slots+numslots && slots+numslots < v.size();
     }
 
     int chooseslot(int flags) const
@@ -121,7 +121,7 @@ static SoundChannel &newchannel(int n, soundslot *slot, const vec *loc = nullptr
         loc = &ent->o;
         ent->flags |= EntFlag_Sound;
     }
-    while(!(static_cast<int>(channels.size()) > n))
+    while(!(channels.size() >n))
     {
         channels.push_back(channels.size());
     }
@@ -142,7 +142,7 @@ static SoundChannel &newchannel(int n, soundslot *slot, const vec *loc = nullptr
 //sets a channel as not being in use
 static void freechannel(int n)
 {
-    if(!(static_cast<int>(channels.size()) > n) || !channels[n].inuse)
+    if(!(channels.size() > n) || !channels[n].inuse)
     {
         return;
     }
@@ -613,7 +613,7 @@ static struct SoundType
     }
     void preloadsound(int n)
     {
-        if(nosound || !(static_cast<int>(configs.size()) > n))
+        if(nosound || !(configs.size() > n))
         {
             return;
         }
@@ -788,7 +788,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
         return -1;
     }
     SoundType &sounds = ent || flags&Music_Map ? mapsounds : gamesounds;
-    if(!(static_cast<int>(sounds.configs.size()) > n)) //sound isn't within index
+    if(!(sounds.configs.size() > n)) //sound isn't within index
     {
         conoutf(Console_Warn, "unregistered sound: %d", n);
         return -1;
@@ -801,7 +801,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
         int rad = radius > 0 ? (maxsoundradius ? std::min(maxsoundradius, radius) : radius) : maxsoundradius;
         if(camera1->o.dist(*loc) > 1.5f*rad)
         {
-            if(static_cast<int>(channels.size()) > chanid && sounds.playing(channels[chanid], config))
+            if(channels.size() > chanid && sounds.playing(channels[chanid], config))
             {
                 Mix_HaltChannel(chanid);
                 freechannel(chanid);
@@ -839,7 +839,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
             return -1;
         }
     }
-    if(static_cast<int>(channels.size()) > chanid)
+    if(channels.size() > chanid)
     {
         SoundChannel &chan = channels[chanid];
         if(sounds.playing(chan, config))
@@ -877,7 +877,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
             break;
         }
     }
-    if(chanid < 0 && static_cast<int>(channels.size()) < maxchannels)
+    if(chanid < 0 && channels.size() < maxchannels)
     {
         chanid = channels.size();
     }
@@ -928,7 +928,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
 
 bool stopsound(int n, int chanid, int fade)
 {
-    if(!(static_cast<int>(gamesounds.configs.size()) > n) || !(static_cast<int>(channels.size()) > chanid) || !gamesounds.playing(channels[chanid], gamesounds.configs[n]))
+    if(!(gamesounds.configs.size() > n) || !(channels.size() > chanid) || !gamesounds.playing(channels[chanid], gamesounds.configs[n]))
     {
         return false;
     }
